OF-2682: Move duplicate close listener management to new abstract Connection class
